% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{add_pes}
\alias{add_pes}
\title{Add partial η² (pes) to standard ANOVA tables}
\usage{
add_pes(results_tbl, model)
}
\arguments{
\item{results_tbl}{Results table with Effect column.}

\item{model}{Fitted model.}
}
\value{
Results table with pes column.
}
\description{
Uses effectsize::eta_squared() to compute partial η² and merge into results.
}
