% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_raw_npx}
\alias{add_raw_npx}
\title{Add raw NPX values to results}
\usage{
add_raw_npx(results, long_df)
}
\arguments{
\item{results}{Results tibble.}

\item{long_df}{Long-format data with NPX/ELISA.}
}
\value{
Results tibble with added sample columns.
}
\description{
Pivot NPX (or ELISA) values wide by SampleID and join onto results.
}
