% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{afex_test}
\alias{afex_test}
\title{Wrapper for afex ANOVA/LMM}
\usage{
afex_test(
  long_df,
  formula,
  treatment,
  control,
  variable = NULL,
  observed = NULL,
  alpha = 0.1,
  report_lfc = FALSE,
  emm = NULL,
  alt_lfc_col = NULL,
  lmer = FALSE,
  rm_incomplete = TRUE
)
}
\value{
List with combined anova and emm_int.
}
\description{
Parallelized wrapper for \code{one_car()} or \code{one_mixed()} across all proteins.
}
