% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{append_cols}
\alias{append_cols}
\title{Append annotation columns to all contrasts}
\usage{
append_cols(res_list, annot_tbl, by)
}
\arguments{
\item{res_list}{Named list of contrast results. Each element must
contain a tibble \code{$anova}.}

\item{annot_tbl}{Tibble or data frame with annotation columns to join.}

\item{by}{Character name(s) of key columns to match between
\code{res_list[[nm]]$anova} and \code{annot_tbl}.}
}
\value{
The same results list, with annotation columns appended
to each \code{$anova} table.
}
\description{
Adds extra annotation columns (e.g. gene IDs) to each contrast’s
\code{$anova} table by performing a left join with a provided
annotation tibble.
}
