% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{append_geneid}
\alias{append_geneid}
\title{Append GeneID annotations}
\usage{
append_geneid(res_list)
}
\arguments{
\item{res_list}{Results list.}
}
\value{
Results list with GeneIDs.
}
\description{
Maps assay symbols to ENSEMBL GeneIDs and appends them to result tables.
}
