% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assess_sample_fit}
\alias{assess_sample_fit}
\title{Assess sample fit with Cook’s distance or residuals}
\usage{
assess_sample_fit(res_list, metric_fun, plot_fun, extra_pkgs = character())
}
\arguments{
\item{res_list}{Results list.}

\item{metric_fun}{Function computing diagnostics.}

\item{plot_fun}{Function plotting diagnostics.}

\item{extra_pkgs}{Required packages.}
}
\value{
Named list of plots.
}
\description{
Generic driver to compute sample diagnostics (Cook’s distance or residuals)
and return plots per contrast.
}
