% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bridgeability}
\alias{bridgeability}
\title{Bridgeability plot}
\usage{
bridgeability(long_data, ...)
}
\arguments{
\item{long_data}{NPX data.}

\item{...}{Extra args passed to \code{olink_bridgeability_plot()}.}
}
\value{
Plotly object.
}
\description{
Wrapper for \code{OlinkAnalyze::olink_bridgeability_plot}, saving a PDF
and returning interactive Plotly.
}
