% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{choose_frameworks}
\alias{choose_frameworks}
\title{Run core statistical tests}
\usage{
choose_frameworks(params)
}
\arguments{
\item{params}{Parameter list from \code{get_params()}.}
}
\value{
Named list of test results (tables).
}
\description{
Executes the specified set of core tests (ANOVA (may contain
configured posthocs), linear mixed models,
t-test, Wilcoxon, Kruskal–Wallis, Friedman, ordinal regression).
}
