% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{enrich_correlations}
\alias{enrich_correlations}
\title{Run Enrichment for Correlation Results}
\usage{
enrich_correlations(res_list, config)
}
\arguments{
\item{res_list}{Named list of correlation tibbles.}

\item{config}{List containing enrichment configuration.}
}
\value{
Nested list: enrich_list[\link{variable}][\link{method}] → tibble of results.
}
\description{
For each metadata variable, runs enrichment analysis with methods specified
in \code{config$enrichment$methods}. Supports \code{"ORA"} and \code{"GSEA"}. Returns a
nested list of results.
}
