% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{get_num_sig_corr}
\alias{get_num_sig_corr}
\title{Get Number of Significant Correlations for One Variable}
\usage{
get_num_sig_corr(spearman_res, variable)
}
\arguments{
\item{spearman_res}{Named list of correlation result tibbles.}

\item{variable}{Metadata variable name.}
}
\value{
Integer count of significant proteins.
}
\description{
Retrieves the number of significant proteins for a single metadata variable
from \code{report_num_sig_corr()}.
}
