% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{interaction_emm}
\alias{interaction_emm}
\title{Compute interaction contrasts via emmeans::emtrends}
\usage{
interaction_emm(mdl, interaction_spec)
}
\arguments{
\item{mdl}{Fitted afex model.}

\item{interaction_spec}{Character like "Factor:Covariate".}
}
\value{
List of \code{trends}, \code{contrasts}.
}
\description{
Compute interaction contrasts via emmeans::emtrends
}
