% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{kruskal_test}
\alias{kruskal_test}
\title{Kruskal-Wallis or Friedman test}
\usage{
kruskal_test(
  long_df,
  variable,
  alpha = 0.05,
  dependence = FALSE,
  subject = NULL,
  verbose = TRUE
)
}
\arguments{
\item{long_df}{Long-format NPX data.}

\item{variable}{Grouping factor.}

\item{alpha}{FDR threshold.}

\item{dependence}{If TRUE run Friedman test with subject ID.}

\item{subject}{Subject column for Friedman.}
}
\value{
Tibble of results with epsilon² or Kendall's W.
}
\description{
Kruskal-Wallis or Friedman test
}
