% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{linear_mixed_model}
\alias{linear_mixed_model}
\title{Linear mixed model via lmerTest}
\usage{
linear_mixed_model(
  long_df,
  variable,
  random,
  covariates = NULL,
  outcome = "NPX",
  alpha = 0.05,
  return.covariates = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{long_df}{Long-format NPX data.}

\item{variable}{Fixed factors.}

\item{random}{Random effect column.}

\item{covariates}{Optional covariates.}

\item{outcome}{Outcome column.}

\item{alpha}{Threshold.}
}
\value{
Tibble of LMM results.
}
\description{
Linear mixed model via lmerTest
}
