% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_priors}
\alias{make_priors}
\title{Make Priors for BRMS}
\usage{
make_priors(df, contrast, sd_treat = 1)
}
\arguments{
\item{df}{Long-format data frame with NPX and covariates.}

\item{contrast}{One of \code{"disease_vs_control"}.}

\item{sd_treat}{SD for treatment effect prior (default 1).}
}
\value{
List of priors (class \code{brmsprior}).
}
\description{
Constructs weakly informative priors for one contrast (e.g., Disease vs control).
Priors are tailored to each contrast
using empirical SDs and kurtosis estimates.
}
