% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mdcat}
\alias{mdcat}
\title{Print Markdown text}
\usage{
mdcat(...)
}
\arguments{
\item{...}{Objects or strings to print.}
}
\value{
No return value; called for side effects.
}
\description{
Simple wrapper around \code{cat()} to print text as Markdown.
Use inside an R chunk with \code{results="asis"} enabled.
}
