% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{more_levels}
\alias{more_levels}
\title{Keep Specific Factor Levels}
\usage{
more_levels(df, column, treatment, control)
}
\arguments{
\item{df}{Data frame.}

\item{column}{Grouping column.}

\item{treatment}{Treatment levels.}

\item{control}{control levels.}
}
\value{
Data frame with filtered factor levels.
}
\description{
Keeps specified treatment and control levels as factor levels without
collapsing them.
}
