% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{one_car}
\alias{one_car}
\title{ANCOVA via afex::aov_car}
\usage{
one_car(
  df_sub,
  key,
  formula,
  form_obj,
  observed,
  variable,
  alt_lfc_col,
  report_lfc,
  emm,
  alpha,
  outcome,
  treatment,
  control
)
}
\value{
List with anova table and emm_int table.
}
\description{
ANCOVA via afex::aov_car
}
