% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{one_mixed}
\alias{one_mixed}
\title{Mixed-effects ANOVA via afex::mixed}
\usage{
one_mixed(
  df_sub,
  key,
  formula,
  form_obj,
  observed,
  variable,
  alt_lfc_col,
  report_lfc,
  emm,
  alpha,
  outcome,
  treatment,
  control
)
}
\value{
List with anova table and emm_int table.
}
\description{
Mixed-effects ANOVA via afex::mixed
}
