% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ordinal_regression}
\alias{ordinal_regression}
\title{Ordinal regression (ordinal ANOVA)}
\usage{
ordinal_regression(
  long_df,
  variable,
  covariates = NULL,
  alpha = 0.05,
  outcome = "NPX",
  return.covariates = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{long_df}{Long-format NPX data.}

\item{variable}{Factor(s).}

\item{covariates}{Covariates.}

\item{alpha}{Threshold.}

\item{outcome}{Ordinal outcome column.}
}
\value{
Tibble with ordinal ANOVA results.
}
\description{
Ordinal regression (ordinal ANOVA)
}
