% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parallel_test_runner}
\alias{parallel_test_runner}
\title{Run all tests in parallel}
\usage{
parallel_test_runner(contrast_list, se_list, config)
}
\arguments{
\item{contrast_list}{Contrast parameters list.}

\item{se_list}{SummarizedExperiment list.}

\item{config}{Config list with \code{parallel} and \code{cores}.}
}
\value{
Results list.
}
\description{
Runs \code{run_tests()} in parallel if enabled in config, otherwise sequentially.
}
