% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_tests}
\alias{run_tests}
\title{Run tests for multiple contrasts}
\usage{
run_tests(contrast_list, se_list)
}
\arguments{
\item{contrast_list}{Named list of contrast specification lists.}

\item{se_list}{Named list of \linkS4class{SummarizedExperiment} objects.}
}
\value{
Named list of results for all contrasts.
}
\description{
Iterates over a list of contrast specifications and runs \code{tests()}
for each.
}
