% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_results}
\alias{sort_results}
\title{Sort results consistently}
\usage{
sort_results(results)
}
\arguments{
\item{results}{Results data frame.}
}
\value{
Tibble sorted by significance and effect size.
}
\description{
Sorts result tables by adjusted p-value and/or effect size.
Special handling for contrasts with multiple \verb{estimate_*} columns
(e.g. therapeutic reversal).
}
