% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{t_test}
\alias{t_test}
\title{Two-sample or paired t-test (Welch)}
\usage{
t_test(long_df, variable, pair_id = NULL, alpha = 0.05, ...)
}
\arguments{
\item{long_df}{Long-format NPX data.}

\item{variable}{Grouping factor (2 levels).}

\item{pair_id}{Optional pairing column.}

\item{alpha}{FDR threshold.}
}
\value{
Tibble of t-test results.
}
\description{
Two-sample or paired t-test (Welch)
}
