% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests}
\alias{tests}
\title{Run all tests for a contrast}
\usage{
tests(contrast_obj, se_list)
}
\arguments{
\item{contrast_obj}{List specifying a contrast.}

\item{se_list}{Named list of \linkS4class{SummarizedExperiment} objects.}
}
\value{
List of test results and parameters for the contrast.
}
\description{
Wrapper to run tests for one contrast, returning
results together with the parameter list.
}
