% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{top_sig}
\alias{top_sig}
\title{Retrieve top significant proteins}
\usage{
top_sig(df)
}
\arguments{
\item{df}{Results data frame with \code{estimate}, \code{Adjusted_pval}, \code{Threshold}.}
}
\value{
List with \code{OlinkID} and \code{Assay} vectors.
}
\description{
Gets top 5 up-regulated and top 5 down-regulated proteins by adjusted
p-value.
}
