% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{two_levels}
\alias{two_levels}
\title{Collapse Groups to Two Levels}
\usage{
two_levels(df, column, treatment, control)
}
\arguments{
\item{df}{Data frame.}

\item{column}{Column to collapse.}

\item{treatment}{Values treated as \code{"treatment"}.}

\item{control}{Values treated as \code{"control"}.}
}
\value{
Data frame with two-level factor.
}
\description{
Collapses a grouping variable into treatment vs control.
}
