% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate contrast parameter object(s)}
\usage{
validate_parameters(params_list, se_list, sensitivity = FALSE)
}
\arguments{
\item{params_list}{Single contrast list or list of contrast lists.}

\item{se_list}{Named list of \linkS4class{SummarizedExperiment} objects}

\item{sensitivity}{Boolean if TRUE, don't worry about the Excel sheet name
length requirement.}
}
\value{
Invisibly returns the original \code{params_list} if validation passes.
}
\description{
Checks that contrast parameter objects are correctly structured, names are
unique, and required fields are present. Also validates factor or numeric
column types against the SummarizedExperiment data.
}
