% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wilcox_test}
\alias{wilcox_test}
\title{Wilcoxon rank-sum or signed-rank test}
\usage{
wilcox_test(long_df, variable, pair_id = NULL, alpha = 0.05, ...)
}
\arguments{
\item{long_df}{Long-format NPX data.}

\item{variable}{Grouping factor (2 levels).}

\item{pair_id}{Optional pairing column.}

\item{alpha}{FDR threshold.}
}
\value{
Tibble of Wilcoxon results.
}
\description{
Wilcoxon rank-sum or signed-rank test
}
